+++ src/components/manage/Toolbar/ToolbarComponents.jsx
@@ -7,7 +7,7 @@ import { defineMessages, injectIntl } from 'react-intl';
 
 import { Icon, Display, Workflow } from '@plone/volto/components';
 import { getBaseUrl } from '@plone/volto/helpers';
-import { settings } from '~/config';
+import config from '@plone/volto/registry';
 
 import addSVG from '@plone/volto/icons/add-document.svg';
 import clearSVG from '@plone/volto/icons/clear.svg';
@@ -18,6 +18,8 @@ import pastanagalogo from '@plone/volto/components/manage/Toolbar/pastanaga.svg'
 import penSVG from '@plone/volto/icons/pen.svg';
 import rightArrowSVG from '@plone/volto/icons/right-key.svg';
 import userSVG from '@plone/volto/icons/user.svg';
+
+// import { matchPath } from 'react-router';
 
 const messages = defineMessages({
   edit: {
@@ -164,6 +166,7 @@ export const ContentsButton = (props) => {
 };
 
 export const AddButton = (props) => {
+  const { settings } = config;
   return (
     props.content &&
     ((props.content.is_folderish && props.types.length > 0) ||
@@ -268,6 +271,7 @@ export const SharingAction = (props) => {
 export const ManageTranslations = (props) => {
   const editAction = find(props.actions.object, { id: 'edit' });
   const path = getBaseUrl(props.pathname);
+  const { settings } = config;
   return (
     editAction &&
     settings.isMultilingual && (
@@ -299,14 +303,10 @@ export const UserButton = (props) => (
 );
 
 export const Bottom = (props) => {
-  const { actionComponents = [] } = props;
   return (
     <>
       <img className="minipastanaga" src={pastanagaSmall} alt="" />
-      {!props.hideDefaultViewButtons &&
-        actionComponents.map((BottomComponent, index) => (
-          <BottomComponent {...props} key={index} />
-        ))}
+      {!props.hideDefaultViewButtons && props.children}
       <div className="divider" />
       <div className="pastanagalogo">
         <img src={pastanagalogo} alt="" />